+++ packages/elements/flag/src/types.js
@@ -2,11 +2,10 @@
 
 /* eslint-disable import/prefer-default-export */
 
-import type { Element, Node } from 'react';
-import type { PropType } from 'babel-plugin-react-flow-props-to-prop-types'; // eslint-disable-line import/no-extraneous-dependencies
+import type { Node } from 'react';
 
-export type ChildrenType = PropType<Array<Element<any>> | Element<any>, any>;
-export type ElementType = PropType<Element<mixed>, any>;
+export type ChildrenType = any;
+export type ElementType = any;
 export type FunctionType = (...args: Array<any>) => mixed;
 export type MouseEventFunctionType = (event: MouseEvent) => any;
 export type GenericEventFunctionType = (event: Event) => any;