+++ packages/fabric/editor-test-helpers/src/mock-extension-data.ts
@@ -17,6 +17,8 @@ export const inlineExtensionData = [
           placeholder: [
             {
               data: {
+                width: 274,
+                height: 30,
                 url:
                   '//pug.jira-dev.com/wiki/plugins/servlet/confluence/placeholder/macro?definition=e3N0YXR1czpzdWJ0bGU9dHJ1ZXxjb2xvdXI9R3JlZW58dGl0bGU9T0t9&locale=en_GB&version=2',
               },
@@ -43,6 +45,32 @@ export const inlineExtensionData = [
           placeholder: [
             {
               data: { url: '' },
+              type: 'icon',
+            },
+          ],
+        },
+      },
+    },
+  },
+  {
+    type: 'inlineExtension' as ExtensionType,
+    attrs: {
+      extensionType: 'com.atlassian.confluence.macro.core',
+      extensionKey: 'status',
+      parameters: {
+        macroParams: {
+          color: { value: 'Grey' },
+          title: { value: 'Medium' },
+          subtle: { value: true },
+        },
+        macroMetadata: {
+          macroId: { value: new Date().valueOf() },
+          placeholder: [
+            {
+              data: {
+                url:
+                  '//pug.jira-dev.com/wiki/download/resources/com.atlassian.confluence.plugins.status-macro/images/status-icon.png',
+              },
               type: 'icon',
             },
           ],
@@ -191,6 +219,10 @@ export const bodiedExtensionData = [
                 type: 'underline',
               },
             ],
+          },
+          {
+            type: 'text',
+            text: ' ',
           },
           {
             type: 'inlineExtension' as ExtensionType,