+++ packages/utils/docs/CHANGELOG.md
@@ -1,5 +1,8 @@
 # @atlaskit/docs
 
+## 2.4.2
+- [patch] Refactor code helper function to fix React re-render bug. [8dcb772](https://bitbucket.org/atlassian/atlaskit-mk-2/commits/8dcb772)
+
 ## 2.4.1
 
 ## 2.4.0